move the function to helper and re-export from export.js
