+++ export.js
@@ -7,21 +7,12 @@ const gulp = require('gulp');
 const helmet = require('helmet');
 const bodyParser = require('body-parser');
 const streamWatcher = require('./src/lib/utils/stream-watcher');
+const { fileWatcher } = require('./src/lib/utils/helper');
 // Re-export
-exports.gulp = gulp;
-exports.helmet = helmet;
-exports.bodyParser = bodyParser;
-exports.streamWatcher = streamWatcher;
-exports.watcher = function(filePaths, callback, verbose = true, debounce = 300) {
-  let files = [];
-  return streamWatcher(filePaths, verbose)
-    .doAction(f => files.push(f))
-    .debounce(debounce)
-    .onValue(() => {
-      if (files.length) {
-        callback(files);
-        // Reset
-        files = [];
-      }
-    });
+module.exports = {
+  gulp,
+  helmet,
+  bodyParser,
+  streamWatcher,
+  fileWatcher
 };
